- Refactored to avoid too many nested logic switches - I still think this isn't easy to maintain, but would need to modify the Item class for that and the goblin won't allow me to - Time spent: 15 minutes
Some thoughts: modifying the Item class, or turning it into an interface, would allow to have several classes, each one handling its own SellIn and Quality values (and anything else). This would prevent unnecesary blocks of nested logic switches, and allow the usage of streaming and lambdas, which in this case would make reading and maintaining the code easier.